requirements.txt
langchain>=0.1
openai>=1.15.0
python-dotenv>=1.0
tqdm>=4.0
langchain_openai
langchain_community
langchain_experimental
langgraph
langchain_google_genai
langchainhub